changes done

+++ lib/actions.ts
@@ -116,7 +116,7 @@ export async function addTrade(formData: FormData) {
   const tradeDate = (formData.get("tradeDate") as string) || new Date().toLocaleString('en-US', { timeZone: 'Asia/Dubai' }).split(',')[0]
   const symbol = "XAUUSD"
   const tradeType = formData.get("tradeType") as string
-  const lotSize = Number.parseInt(formData.get("lotSize") as string)
+  const lotSize = Number.parseFloat(formData.get("lotSize") as string)
   const entryPrice = Number.parseFloat(formData.get("entryPrice") as string)
   const exitPrice = Number.parseFloat(formData.get("exitPrice") as string) // required
   const notes = formData.get("notes") as string
